Add original arduino file
this is unfortunately the old version which does not have a smooth
acceleration of the servo motor. The improved version is gone.
Only use it as a reference and do not upload it to the ESP32 mounted on the barbot!
For testing, use a different ESP32 board and clip it to the barbot.